# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export type MapDataPayloadCountryDetails = { [key: string]: unknown };

export type MapDataPayloadStateDetails = { [key: string]: unknown };

export type MapDataPayloadProvinceDetails = { [key: string]: unknown };

export type MapDataPayloadStadiumDetails = { [key: string]: unknown };

export type MapDataPayloadParkDetails = { [key: string]: unknown };

export type MapDataPayloadTccDetails = { [key: string]: unknown };

export type MapDataPayloadNotesByKey = { [key: string]: string };

export interface MapDataPayload {
  /** Schema version for forward-compatibility tracking. */
  schemaVersion?: number;
  visitedCountries?: string[];
  visitedStates?: string[];
  visitedProvinces?: string[];
  visitedStadiums?: string[];
  visitedParks?: string[];
  tccVisited?: string[];
  bucketCountries?: string[];
  bucketStates?: string[];
  bucketProvinces?: string[];
  bucketStadiums?: string[];
  bucketParks?: string[];
  tccBucket?: string[];
  countryDetails?: MapDataPayloadCountryDetails;
  stateDetails?: MapDataPayloadStateDetails;
  provinceDetails?: MapDataPayloadProvinceDetails;
  stadiumDetails?: MapDataPayloadStadiumDetails;
  parkDetails?: MapDataPayloadParkDetails;
  tccDetails?: MapDataPayloadTccDetails;
  notesByKey?: MapDataPayloadNotesByKey;
  profileName?: string;
}

export interface MapDataEnvelope {
  data: MapDataPayload | null;
}

export interface SaveMapDataResponse {
  ok: boolean;
}

export interface PhotoRecord {
  /** UUID of the photo record. */
  id: string;
  /** Relative URL to fetch the photo content, e.g. /api/photos/<id>/content. */
  url: string;
  /** @maxLength 120 */
  caption: string;
  /**
   * @minimum 0
   * @maximum 2
   */
  position: number;
  createdAt: string;
}

export interface PhotoListResponse {
  photos: PhotoRecord[];
}

export interface UpdatePhotoBody {
  /** @maxLength 120 */
  caption?: string;
}

export interface DeletePhotoResponse {
  ok: boolean;
}

export interface UserProfile {
  userId: string;
  username: string;
  displayName?: string | null;
  usernameSet: boolean;
}

export interface SetUsernameBody {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
}

export interface SetUsernameResponse {
  ok: boolean;
  username: string;
  usernameSet: boolean;
}

export interface UserSearchEntry {
  userId: string;
  username: string;
  displayName?: string | null;
}

export interface UserSearchResponse {
  users: UserSearchEntry[];
}

export interface ConnectionOtherUser {
  userId: string;
  username?: string | null;
  displayName?: string | null;
}

export type ConnectionRecordStatus =
  (typeof ConnectionRecordStatus)[keyof typeof ConnectionRecordStatus];

export const ConnectionRecordStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
} as const;

export type ConnectionRecordDirection =
  (typeof ConnectionRecordDirection)[keyof typeof ConnectionRecordDirection];

export const ConnectionRecordDirection = {
  incoming: "incoming",
  outgoing: "outgoing",
} as const;

export interface ConnectionRecord {
  id: string;
  status: ConnectionRecordStatus;
  direction: ConnectionRecordDirection;
  createdAt: string;
  respondedAt?: string | null;
  otherUser?: ConnectionOtherUser;
}

export type ConnectionsResponsePending = {
  incoming: ConnectionRecord[];
  outgoing: ConnectionRecord[];
};

export interface ConnectionsResponse {
  pending: ConnectionsResponsePending;
  accepted: ConnectionRecord[];
}

export interface DestinationRow {
  userId: string;
  category: string;
  destinationId: string;
  isVisited: boolean;
  isBucket: boolean;
  firstVisitedYear?: number | null;
  lastVisitedYear?: number | null;
  timesVisited?: number | null;
}

export interface CompareUserData {
  userId: string;
  username?: string | null;
  displayName?: string | null;
  destinations: DestinationRow[];
}

export interface CompareResponse {
  me: CompareUserData;
  other: CompareUserData;
}

export interface LeaderboardVisitedCounts {
  country?: number;
  us_state?: number;
  ca_province?: number;
  tcc?: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  username?: string | null;
  displayName?: string | null;
  isMe: boolean;
  visitedCounts: LeaderboardVisitedCounts;
  totalVisited: number;
}

export interface LeaderboardResponse {
  leaderboard: LeaderboardEntry[];
}

export interface DestinationStatEntry {
  destinationId: string;
  visitedCount: number;
  bucketCount: number;
}

export interface AggregateStatsResponse {
  category: string;
  /** Total distinct users in the system. */
  totalUsers: number;
  destinations: DestinationStatEntry[];
}

export interface DestinationStatsResponse {
  category: string;
  destinationId: string;
  visitedCount: number;
  bucketCount: number;
  totalUsers: number;
  /** Percentage of total users who have visited (0–100). */
  visitedPct: number;
  /** Percentage of total users with this on their bucket list (0–100). */
  bucketPct: number;
}

export type ListPhotosParams = {
  /**
   * Destination category (country, state, province, tcc, etc.)
   */
  category: string;
  /**
   * Destination identifier within its category.
   */
  destinationId: string;
};

export type UploadPhotoBody = {
  file: Blob;
  category: string;
  destinationId: string;
  /**
   * @minimum 0
   * @maximum 2
   */
  position?: number;
  /** @maxLength 120 */
  caption?: string;
};

export type SearchUsersParams = {
  /**
   * @minLength 2
   * @maxLength 30
   */
  q: string;
};

export type GetAggregateStatsParams = {
  category: GetAggregateStatsCategory;
  /**
   * @minimum 1
   * @maximum 200
   */
  limit?: number;
};

export type GetAggregateStatsCategory =
  (typeof GetAggregateStatsCategory)[keyof typeof GetAggregateStatsCategory];

export const GetAggregateStatsCategory = {
  country: "country",
  us_state: "us_state",
  ca_province: "ca_province",
  tcc: "tcc",
} as const;

export type GetDestinationStatsParams = {
  category?: string;
};