# lib/api-zod/src/generated/types/connectionRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectionOtherUser } from "./connectionOtherUser";
import type { ConnectionRecordDirection } from "./connectionRecordDirection";
import type { ConnectionRecordStatus } from "./connectionRecordStatus";

export interface ConnectionRecord {
  id: string;
  status: ConnectionRecordStatus;
  direction: ConnectionRecordDirection;
  createdAt: Date;
  respondedAt?: Date | null;
  otherUser?: ConnectionOtherUser;
}