# lib/api-zod/src/generated/types/destinationRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DestinationRow {
  userId: string;
  category: string;
  destinationId: string;
  isVisited: boolean;
  isBucket: boolean;
  firstVisitedYear?: number | null;
  lastVisitedYear?: number | null;
  timesVisited?: number | null;
}